useFirebase and AuthContext
previously in useFirebase hook was creating problem so I removed prevoius one and rewrite the code and also rewrote AuthContext with UserContext and now there is no problem. Previously problem was because I initialized firebase in App.js but wrote getAuth() in useFirebase hook, this was conflicting. Now I've initialized firebase in useFirebase hook.